Q2_Custom_Embedder/requirements.txt
pypdf>=4.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
sentence-transformers>=2.2.0
torch>=2.0.0
pandas>=2.0.0
seaborn>=0.12.0
umap-learn>=0.5.0
hdbscan>=0.8.0